removed unused system text json deps, cleaned up
