+++ KestrelMock/KestrelMock.csproj
@@ -13,7 +13,6 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
 		<PackageReference Include="Microsoft.AspNetCore.Hosting" Version="2.2.7" />
-		<PackageReference Include="System.Text.Json" Version="4.7.2" />
 	</ItemGroup>
 
 </Project>
